+++ src/main/java/com/dace/dmgr/combat/character/Controller.java
@@ -65,7 +65,7 @@ public abstract class Controller extends Character {
 
     @Override
     @MustBeInvokedByOverriders
-    public void onDamage(@NonNull CombatUser victim, @Nullable Attacker attacker, int damage, @NonNull DamageType damageType, Location location, boolean isCrit) {
+    public void onDamage(@NonNull CombatUser victim, @Nullable Attacker attacker, double damage, @NonNull DamageType damageType, Location location, boolean isCrit) {
         CooldownUtil.setCooldown(victim, HEAL_COOLDOWN_ID, RoleTrait2Info.ACTIVATE_DURATION);
     }
